Painful Dialogue Learning
Added node system using the following scripts:

- Actor.cs
- Dialogue.cs
- DialogueManager.cs
- DialogueNode.cs
- DialogueResponse.cs

These are all in the Dialogue folder I created within Assets. With them, there is a prefab for selecting multiple choice responses to dialogue. This is called depending on how many responses there are in the dialogue. 2 character dialogues have been used for testing. Rebel1 and Rebel2. The first was the main lab rat during the construction of the dialogue. My main mission with Rebel2 dialogue was to jump to it as soon as Rebel1 dialogue had ended and called for it. I've failed to do this so I'm just uploading my stuff for now.

Jumping to another character in the middle of one's dialogue does not seem worth the effort. I would recommend combining Rebel1 and Rebel2's dialogue in the dialogue key. Either this or someone who isn't fumbling around unity like an idiot using a language they've pretty much never touched might be able to come up with a solution without breaking a sweat. Anyways this is all I have for dialogue for now.